Fix return type of VL53L0X milliseconds()

+++ VL53L0X.cpp
@@ -37,7 +37,7 @@ using namespace LinuxDuino;
 uint64_t milliseconds() {
 	timespec ts;
 	clock_gettime(CLOCK_MONOTONIC, &ts);
-	return (ts.tv_sec * 1000 + ts.tv_nsec / 1000000);
+    return (uint64_t) (ts.tv_sec * 1000 + ts.tv_nsec / 1000000);
 }
 
 /*** Constructors ***/
